fixed cors, added bg color transition, added experiences
fixed cors by removing the allowOrigin, because it was set to localhost and the server started at 127.0.0.1 or something like that, it's the same but it's not. Added the background color transition so when the user changes between light and dark mode it's smooth. Added experiences, it's definitely not what i want but it's somethig. Will improve it.