refactor: added DTOs and Services, request mapping before controller
added DTOs (Data Transfer Objects) and services for every class so it's possible to partially update an object using a PATCH request. Previously, I tried to do it with multiple PUT endpoints to every attribute for each class, which was a mess and didn't work as it should. Now, the Controller communicates with the Service and the Service with the DTO and the Repository. The class endpoint now is set with a Request Mapping before the class so that it isn't needed to write the same thing a lot of times, for example: instead of @GetMapping('/projects/{name}') it is now @GetMapping('/{name}')